// src/app/interfaces/job.model.ts
export interface Job {
  JobId: number;
  Experience:String;
  JobTitle: string;
  Company: string;
  CompanySize:number;
  location: string;
  JobPostingDate: string;
  JobDescription: string;
  duration: string;
  skills: string;
  WorkType: string;
  SalaryRange: string;
  Country: string;
  Category: string;
  DocumentFileName: string;
  // Add other attributes as needed
}